[util] Consolidate custom hooks into one util file
