+++ src/util/hooks.ts
@@ -1,4 +1,4 @@
-import type { NS } from 'netscript';
+import type { NS, UserInterfaceTheme } from 'netscript';
 
 import { makeFuid } from 'util/fuid';
 
@@ -34,4 +34,19 @@ export function useNsUpdate<T>(
     }, [ns, interval, updateFn]);
 
     return data;
+}
+
+/**
+ * Keep a UserInterfaceTheme updated by polling `ns.ui.getTheme()`.
+ *
+ * @param ns - Netscript API instance
+ * @param interval - Milliseconds between theme refreshes
+ * @returns The current theme from the UI
+ */
+export function useTheme(ns: NS, interval = 200): UserInterfaceTheme {
+    return useNsUpdate(ns, interval, getTheme);
+}
+
+function getTheme(ns: NS): UserInterfaceTheme {
+    return ns.ui.getTheme();
 }
